+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/membership/RoomMemberEventHandler.kt
@@ -22,9 +22,29 @@ import org.matrix.android.sdk.api.session.events.model.toModel
 import org.matrix.android.sdk.api.session.room.model.RoomMemberContent
 import org.matrix.android.sdk.internal.session.user.UserEntityFactory
 import io.realm.Realm
+import io.realm.RealmConfiguration
+import io.realm.kotlin.where
+import org.matrix.android.sdk.api.MatrixCallback
+import org.matrix.android.sdk.api.extensions.orTrue
+import org.matrix.android.sdk.api.util.JsonDict
+import org.matrix.android.sdk.internal.database.model.UserEntity
+import org.matrix.android.sdk.internal.database.model.UserEntityFields
+import org.matrix.android.sdk.internal.di.SessionDatabase
+import org.matrix.android.sdk.internal.di.UserId
+import org.matrix.android.sdk.internal.session.profile.GetProfileInfoTask
+import org.matrix.android.sdk.internal.task.TaskExecutor
+import org.matrix.android.sdk.internal.task.configureWith
+import timber.log.Timber
 import javax.inject.Inject
 
-internal class RoomMemberEventHandler @Inject constructor() {
+internal class RoomMemberEventHandler @Inject constructor(
+        @UserId
+        private val userId: String,
+        @SessionDatabase
+        private val realmConfiguration: RealmConfiguration,
+        private val taskExecutor: TaskExecutor,
+        private val getProfileInfoTask: GetProfileInfoTask
+) {
 
     fun handle(realm: Realm, roomId: String, event: Event): Boolean {
         if (event.type != EventType.STATE_ROOM_MEMBER) {
@@ -41,10 +61,54 @@ internal class RoomMemberEventHandler @Inject constructor() {
         }
         val roomMemberEntity = RoomMemberEntityFactory.create(roomId, userId, roomMember)
         realm.insertOrUpdate(roomMemberEntity)
-        if (roomMember.membership.isActive()) {
-            val userEntity = UserEntityFactory.create(userId, roomMember)
-            realm.insertOrUpdate(userEntity)
+
+        if (roomMember.membership.isActive() && shouldUpdateUserEntity(realm, userId, roomMember)) {
+            updateUserEntity(realm, userId, roomMember)
+            if (userId == this.userId) {
+                fetchUserProfile(userId) // To fix #1715 (myroomnick changes the global name)
+            }
         }
         return true
+    }
+
+    private fun shouldUpdateUserEntity(realm: Realm, userId: String, roomMember: RoomMemberContent?): Boolean {
+        return realm
+                .where<UserEntity>()
+                .equalTo(UserEntityFields.USER_ID, userId)
+                .findFirst()
+                ?.let {
+                    it.displayName != roomMember?.displayName
+                            || it.avatarUrl != roomMember.avatarUrl
+                }
+                .orTrue()
+    }
+
+    private fun updateUserEntity(realm: Realm, userId: String, roomMember: RoomMemberContent) {
+        val userEntity = UserEntityFactory.create(userId, roomMember)
+        realm.insertOrUpdate(userEntity)
+    }
+
+    private fun fetchUserProfile(userId: String) {
+        val params = GetProfileInfoTask.Params(userId)
+        getProfileInfoTask
+                .configureWith(params) {
+                    this.callback = object : MatrixCallback<JsonDict> {
+                        override fun onSuccess(data: JsonDict) {
+                            val displayName = data["displayname"] as? String
+                            val avatarUrl = data["avatar_url"] as? String
+                            val userEntity = UserEntity(userId, displayName ?: "", avatarUrl ?: "")
+                            Realm.getInstance(realmConfiguration).use { realm ->
+                                realm.executeTransaction {
+                                    it.insertOrUpdate(userEntity)
+                                }
+                            }
+                        }
+
+                        override fun onFailure(failure: Throwable) {
+                            Timber.e(failure, "Couldn't fetch user profile")
+                        }
+                    }
+                }
+                .executeBy(taskExecutor)
     }
 }